Restructured Client component code
