+++ client/src/components/questions-answers-module/question-list/questionsList.jsx
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import axios from 'axios';
 import Card from '@mui/material/Card';
 
-import Question from './Question.jsx'
+import Question from '../question/question.jsx'
